Made |Transform<class C, int T> *transform| a member of |class Shape|. Deriving |class Point|, |class Path| and |class Solid| from |Shape|.
|Transform<C, T> *transform| is NULL by default in |class Point| objects.
A |Transform| can be allocated by calling |Shape::init_transform|.

The default constructors of |Point|, |Path| and |Solid| have an optional
argument |bool allocate_transform|.  It's default value is |false| in the
|Point| constructor and |true| in the others.

Previously, every |Point| had its own |Transform|.  For |Paths| or |Solids|
where only one |Transform| was used to transform the |Points|, this was a waste
of memory.  By using a pointer, I can use a single |Transform| for a |Path|
or |Solid|.  If the object is changed in such a way that one or more |Points|
need their own |Transforms|, i.e., if a transformation is applied to a part of
an object, then |Transforms| can be allocated for the |Points| that need them.